csi: Report actual thin size for NodeGetVolumeStats
When online resize is taking place, a given node may still see the
smaller size of a device-mapper exposing the former size, before
kubernetes has had the chance to perform the reconcile loop to expand
that device-mapper.  Although it is extremely hard in normal operation
to hit the window where the CSI call might report a smaller size than
the cluster would observe, the more important reason to do this is so
that if we ever get stuck where device-mapper doesn't behave as
intended, the cluster can diagnose that our size reports are
inconsistent between controller and node.

Testing this will not be possible until the rest of volume node
expansion is implemented, and even then, the easiest way to test it is
to hack in an artificial sleep to volume_node_manager to intentionally
delay resizing the device-mapper for several seconds after the
ThinPoolLv has resized.

Note that ControllerGetVolume does NOT get the same treatment; there,
we want to list the offline size of the underlying LV, regardless of
whether the LV is staged.

Signed-off-by: Eric Blake <eblake@redhat.com>