snapshot: Fix deletion hang
While trying to test csi-sanity failures, I reliably hit a hang where
deleting a snapshot got wedged.  I traced it to the following:

If a child object has more than one OwnerReference (true when the
Volume continues to exist and we are trying to delete the Snapshot),
it looks like our reconciler will initially run while our
OwnerReference is still in place; but kubernetes does NOT care whether
we have an active reconciler, it will eventually delete the
OwnerReference from the child object asynchronously to our execution.
If we do something lengthy, like return WatchPending because we are
waiting for the ThinPoolLv reconciler to remove the ThinLv, it only
took a couple of iterations and less than a second before the next
entry into the Snapshot reconciler can see the ThinPoolLv has already
lost the OwnerReference back to the Snapshot.  And once that happens,
the controller loop will no longer wake up Snapshot to do another
Reconcile (basically, the Owns() trigger that wakes us up when
ThinPoolLv changes only works as long as ThinPoolLv still has an
OwnerReference).

To fix it, any time we know we are waiting for a child object to
change but where we are already reconciling a deletion, we NEED to
ensure that we keep polling the child object until it has the state we
want, rather than relying on automatic wakeup due to changes in the
child object. The easiest way to do that is to return an error to the
reconciler, which will then retry us after exponential backoff.  More
elegant might be to explicitly request another reconcile on the same
object after a fixed elapsed time (by using ctrl.Result{RequeueAfter:
time.Second} or similar), but that requires more code just to silence
a few errors in the logs.

While at it, ensure that both Volume and Snapshot use the same
paradigm of requesting that whoever is last to be deleted blocks owner
deletion.

Note that returning success in place of WatchPending when we are
waiting on our OWN object (such as waiting for
Volume.Status.AttachedToNodes to clear) is still okay - the .For()
triggers of the reconciler still work, it is only the .Owns() triggers
that can disappear once DeletionTimestamp is set.  But WatchPending
doesn't tell us which object we are waiting on.

Signed-off-by: Eric Blake <eblake@redhat.com>