blobs: Thin provision starting from 1 GiB
Now that LVM automatic extension is enabled again we can provision
thin-pools with just enough space to get started and let them grow as
data is written to them.

Note that the code previously added 1% for metadata but this was
unnecessary because LVM automatically adds space for metadata on top of
the requested amount.

I have tested that creating lots of sub-1 GiB volumes still works and
that they grow after snapshot creation and overwrite.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>