snapshot: Consolidate RemoveBlob and RemoveSnapshot into one
By changing the signature of NewThinBlobManager to take the pool name
up front, it is possible to reuse a single function for deleting any
ThinLV (both Volume and Snapshot), rather than needing a one-off
wrapper for deleting snapshots.  A future patch will do a similar
cleanup for getting a blob's size.

Note that all callers have access to a pool name, while only the
volume caller has access to a vgName; but since the linear blob
manager also needs constant access to a vgName (and does not have the
luxury that the thin manager has of looking it up from ThinPoolLv), it
is not worth trying to use a factory pattern such as
NewThinBlobManager(ctx, name, pool).WithVGName(vg) just to avoid the
vgName hack in the snapshot caller.

Signed-off-by: Eric Blake <eblake@redhat.com>