snapshot: Simplify blob manager size queries
Upcoming patches for volume expansion will want to query blob size for
volumes, which reuses much of the same code as already existing for
thin snapshots.  Removing an extra parameter (that is now gained from
the thin constructor) makes the reuse for expansion easier.

Note that in isolation, this patch temporarily regresses by reporting
0 size for linear volumes.  The CSI Spec technically allows this
(thanks to things like a driver exposing filesystem mounts on NFS,
where the amount of free space is indeterminate because it depends not
only the server, but on how much other clients are using at the same
time), but the next patch restores the behavior we want.

Signed-off-by: Eric Blake <eblake@redhat.com>