snapshot: Add more CSI sanity validations
Make sure that we are only attempting snapshots of thin blocks (we can
probably also do thin filesystems, but that would be better if we
freeze and thaw the filesystem around the snapshot creation).  Also,
the CSI spec requires us to diagnose attempts to create a snapshot
with an invalid source, or to recreate a snapshot with a different
source.

Signed-off-by: Eric Blake <eblake@redhat.com>