volume: Query actual lvm size
Instead of blindly copying spec to status, we should grab actual size
from the lvm volume (Linear reads it directly in Volume, while Thin
has a layer of indirection through ThinPoolLv.Status.ThinLvs[]).  This
will become important in upcoming patches that perform volume
expansion - a resize isn't complete until we can read back the new
size from the device.

The placement of the size check after a successful CreateBlob() is
intentional - we should only update size exposed to CSI when
conditions["Available"] is true (the CSI spec says that a second
Create call while the first is still in progress must not change the
size, and we don't have to worry about odd interactions between a
blkdiscard process started at one size with an expansion to a
different size - but we do intend to use the documented CreateBlob
ability to resize a linear blob).

Similarly, when sizing device mapper wrappers, we should use the size
of the LV as it currently is, rather than a potentially larger
requested size that is still pending expansion.

Signed-off-by: Eric Blake <eblake@redhat.com>