snapshot: Fix sanity checking of parameters
Our deployment asked CSI to hand us several "csi.storage.k8s.io/*"
extra parameters on top of what is present in the VolumeSnapshotClass,
which then caused tests/snapshot.sh to fail when len(req.Parameters)
was non-zero.  But since we aren't actually depending on those
parameters, let's just change the deployment to quit asking for them.

Do the same for volumes and StorageClass parameter synthesis.

(It doesn't help that we are still battling a case of gitlab merging
commits before Jenkins has had a chance to finish regression testing -
because otherwise this would have failed to merge earlier.)

Fixes: commit 9051f5cf ("snapshot: Add more CSI sanity validations")
Signed-off-by: Eric Blake <eblake@redhat.com>