docs: Mention workaround for lvm < 2.03.25
Older lvm has a bug[1] where vgimportdevices hangs if it can't access
the global lock (a problem if the global lock happens to live in the
shared VG being imported).  Although Fedora 40 (the image used by
kcli) has new-enough lvm2, as of this commit, OCP using RHEL CoreOS
from RHEL 9.4 still has only lvm2 2.03.23, so we should document the
obvious workaround of not using the global lock during import, since
that's precisely what lvm 2.03.25 does automatically.

What's more, minikube (based on buildroot) is not only using the older
lvm2 2.03.15, but that distro doesn't even provide /etc/lvm/devices in
its bare-bones install, which was causing __create_ksan_shared_vg to
fail with ENOENT during the touch.  This is also an easy thing to work
around.  As before, keep the docs and test setup in sync.

[1] https://gitlab.com/lvmteam/lvm2/-/commit/a8b8e1f074598

Signed-off-by: Eric Blake <eblake@redhat.com>