dm: Use linear instead of multipath for upper layer
My original understanding of why we needed a multipath layer was
slightly off.  Benjamin Marzinski corrected me: from the application's
point of view, both 'dmsetup suspend' and 'dmsetup message 0 fail_path
/path/to/lower' have the same impact: I/O is queued in the kernel
until a path re-appears.  But with multipath, the queue requires
duplication in the kernel (even though we only set up one backend),
whereas a linear device queues the original kernel bios.

Rather, the real reason we need two layers is that we NEED a way to
close the underlying device, and dmsestup doesn't do that until there
is NO usage of that device (either from a live table installed on the
current layer, or from any other layers using it).  With two layers,
upper fail_path plus lower suspend was enough to get the LV device
untethered so we could deactivate it, whereas a single layer linear
alone would not untether the LV device until a new table is loaded.
And reloading in place with just one layer is not possible (the next
real device to load isn't available until another node starts an NBD
server).  But it is also possible to load a dummy device (zero or
error) in the lower layer to release the live table, without relying
on the fail_path of the upper layer; and as long as the upper layer is
suspended, it doesn't matter what dummy device was loaded in the
interim in the lower layer.

The other benefit of this patch: by not using multipath, we are no
longer stuck with multipath's queue_if_no_path setting that hides I/O
errors rather than percolating them up to the end user.  Testing
showed that a linear volume was able to pause a kubevirt guest when
the LUN was disabled, but that a thin volume was hanging (the LUN
failure was getting queued in the multipath layer rather than reaching
virtiofs to pause the kubevirt guest).  With a linear rather than
multipath volume as our upper layer, I/O errors are now propagated up
for thin volumes just like they are for linear.

This is (yet another) ABI breaking change.  Pre-existing staged thin
volumes created in v0.9.0 will NOT work when upgrading to v0.10.0
because of the name change to the devices being managed.  Ah, the joys
of being in alpha!  Directly backporting this patch to stable-0.9
would not be advisable without some additional back-compat glue to
gracefully handle pre-existing staged volumes.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/104
Signed-off-by: Eric Blake <eblake@redhat.com>