dm: Assign uuids to each dm
Dave Tiegland recommended that all device-mapper devices should be
assigned a uuid at creation (you can rename the device, but not munge
the uuid, once one is assigned).  And by convention, the device-mapper
uuid should begin with a prefix that makes it easy to find all
mappings created by a common creator (for example, 'dmsetup info' on
my laptop finds a CRYPT-LUKS2- and an LVM- prefix).  This allows for
things like 'dmsetup remove -S "uuid=~kubesan"' to quickly remove all
stale device mappings left over if kubesan fails to clean up after
itself.

Having everything in a prefix, rather than both a prefix and suffix,
is also slightly easier to work with, although the rest of the code
base really doesn't care.

Note that this is a uuid (anything universally unique, which really
boils down to a unique namespacing prefix and then that program
guaranteeing uniqueness in its own suffixes) rather than a UUID (in
the RFC4122 sense of the word); we happen to have a UUID embedded in
our uuid when using kubernetes names (pvc-UUID), but not when
processing csi-sanity inputs (where we actually end up with a
redundant-looking kubesan-upper-kubesan-HASH), but no one should
really care.

Suggested-by: Dave Tiegland
Signed-off-by: Eric Blake <eblake@redhat.com>