volume: Implement offline volume resize in cluster reconcile
Previous patches already made sure CreateBlob can perform expansion,
but nothing actually calls CreateBlob once the volume is initially
created and Conditions["Available"] set to true.  Time to fix that.
Add a couple more polymorphic functions to BlobManager to help guide
the logic.

For Linear volumes, CreateBlob is not safe to call unless
Conditions["Available"] is set back to false, but we can't change that
until the volume is offline (not activated or scheduled to be
activated on any node).

For Thin volumes, CreateBlob is safe to call at any time, but setting
Status.SizeBytes should only be done if the volume is offline (if it
is online, the next patch to node reconcile will be handling it
instead), so that we don't have multiple reconcilers competing to make
updates.

This is the first time we actually set Conditions["Abnormal"] - when
someone manually tweaks the Volume.Spec of a filesystem volume (the
intent is that the CSI code will pre-screen and error out rather than
updating filesystem Volumes like that), which may prove useful in
demonstrating the recent patches to add health monitoring.  And until
a later patch actually wires up CSI code to change
Volume.Spec.SizeBytes, for now a resize can only be triggered by
manual `kubectl edit -n kubesan-system volume $vol`.

Signed-off-by: Eric Blake <eblake@redhat.com>