manager: Make WatchPending more useful in logging
Consolidating where we check for WatchPending (such as commit
c4679b81) to turn it into a successful return after short-circuiting a
reconcile had one drawback - it lost details about what was being
waited on.  This adds the ability to document why we are returning
WatchPending, so that the central collection can once again make the
logs give us useful information.

Signed-off-by: Eric Blake <eblake@redhat.com>