manager: Implement cloning from snapshots and volumes
The Volume cluster controller uses the new BlobManager cloning API to
activate source and target blobs on a node.

The Volume node controller then runs a dd(1) worker to perform the data
copy.

Afterwards the Volume cluster controller deactivates the blobs and sets
the Available condition on the new Volume.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>