crd: Snapshot.Spec needs vgName
In order to use a Snapshot as a volume source during volume creation,
that snapshot must reside in a Volume Group that is equally accessible
(thinking ahead to when we support volume groups with sub-cluster
topology).  It's better to be strict up front (require a vgName match)
and relax things later when we start implementing sub-cluster support,
although this particular patch doesn't change anything, a later patch
in the series will tackle that.  But more importantly, in order to
easily check that the user's request is valid (whether it is by vgName
match or by comparing vg topologies), Snapshot should remember its
volume group directly, since the Volume that the Snaphot was created
from may have been deleted in the meantime, and the CSI layer should
not be mucking around with ThinPoolLv objects as the only other source
of truth pre-patch on which VG things are in.

Reverts the removal of this field in b0cf4c45, and further improves
the creation code to actually set the field correctly.

Signed-off-by: Eric Blake <eblake@redhat.com>