crd: Enforce more naming conventions
When doing CEL validation, the Pattern regex limiter is NOT anchored
by default; we had some patterns that were not actually protecting us
from bogus strings because the user only had to have a substring
matching the pattern.  Also, documenting MaxLength on every string
makes it easier for CEL validation to not complain about potentially
over-budget computation when it validates the amount of work it
guesses it will need to do on all our regex.  In fact, CEL validation
didn't like that we had an unbounded number of ThinLV[] entries in
both Spec and Status of ThinPoolLv; once there is a regex validation
on each entry of that array, the array needs to be bounded.  For now,
I picked 64k entries (far more LVs than we ever expect to support in
an entire VG, let alone in one thin pool).

lvm has a combined limit of 128 alphanumeric or -+_ bytes for VG/LV
names, along with additional restrictions such as no leading hyphen
and certain forbidden substrings.  Doing CEL validation of the
substrings is pointless, but we can at least ensure that no one will
try to create a CRD where the names will be too long by restricting VG
and LV names to 63 bytes each (our LV names are under our control, and
a user creating a VG name that long is crazy).  By validating our CRDs
up front, we are also less likely to allow an intentionally malicious
user to hand-write a StorageClass with an invalid VG name in an
attempt to try to get us to execute a shell escape, because we are at
least now assured that there are no shell magic characters in any
vgName field of any of our resources.

Other strings worth validating: We know node names will be RFC 1123
subdomains (253 bytes or less).  We also know that all our object
names that we are willing to work with will be RFC 1123 labels (63
bytes or less - see commit internal/csi/common/validate/ in commit
6297316a).  Right now, there's a slight hole - nothing prevents a user
from manually creating a Volume or Snapshot longer than an RFC 1123
label, even though we then refuse to create an LV by that name; but
kubebuilder does not have an easy way to propagate metadata.name
restrictions through to our CRD yaml even though OpenAPI would let us
add such a restriction.

Signed-off-by: Eric Blake <eblake@redhat.com>