crd: Add ratchet to Status updates
We already have Status.ObservedGeneration (see commit 390b31a5), but
we can put it to slightly better use.  If there are two parallel
reconcilers, and one starts a long-running task while seeing
generation 1 of the CR, another change is made, and the second
reconciler is able to make a quick change based on generation 2 of the
CR, then the second reconciler will write Status.ObservedGeneration to
2 along with its other changes, only to have the first reconciler undo
things back to ObservedGeneration 1.  Unlike Spec changes (where
kubernetes automatically returns a conflict error if things aren't
moving generation forward), we are not protected from this unless we
do it ourselves.  Ideally, our reconcilers are idempotent, and later
ones aren't undoing earlier ones, but a CEL validator helps enforce
that.  This doesn't solve everything (two reconcilers acting on the
same source can still stomp on each other), but it does reduce the
worst impacts of going backwards in time.

Signed-off-by: Eric Blake <eblake@redhat.com>