crd: Update pretty-print columns
Now that Linear volumes resize by going offline, reporting the time
since the volume was first online can be different than the time it
was last resized, so it is worth a second timestamp column.
Similarly, for ThinPoolLv, it is nice to know how old the pool is in
general in addition to how long since it was last activated or
deactivated.

Signed-off-by: Eric Blake <eblake@redhat.com>