tests: Demonstrate use of quotas
If the user is not taking advantage of kubernetes quotas, there is
nothing stopping them from requesting the creation of a thin volume
that far exceeds the size of the volume group.  The creation will
succeed, and as long as the volume stays sparse, they are okay, but
trying to use that much storage can starve all other tenants of the
cluster; and CSI has no way to shrink a volume once it has reported a
larger size.

Kubernetes already has the ability to set defaults (LimitRange) and
quotas (ResourceQuota) on storage claims, so we don't need to
implement anything in our StorageClass to add extra limits.  However,
having a unit test that demonstrates a quota in action could be
helpful to a cluster administrator wanting to ensure no tenant starves
out everyone else.

Signed-off-by: Eric Blake <eblake@redhat.com>