csi: Fast-fail validations of creation from contents
Reject things that we know are impossible or that we don't support
yet, rather than waiting around forever for the manager to make
progress where it cannot do so.  In some cases, it's better to be
strict now where we can relax things later (for example, allowing
cross-VG clones even if we have to use dd may be desirable, as doing
it with a single CSI driver is still easier for the user than having
them spin up an external solution). But even there, once we start
supporting node topologies where VGs are only on a subset of nodes, we
have to be careful to only support clones when both source and
destination can be staged on the same node.

Note: We may decide to do all these validations only in the manager
layer, and instead rely on Conditions["Error"] to let us report those
failures at the gRPC layer.  But for now this was a simple patch that
makes it easier to ensure creation from contents doesn't hang
csi-sanity.

Signed-off-by: Eric Blake <eblake@redhat.com>