volume: Let CreateBlob() expand thin volumes
We already documented that CreateBlob is supposed to allow volume
expansion; actually implement that by passing updated sizes to
ThinPoolLv in the right places.

Now, consider what happens in the code before this patch, given the
following race of events:

volume is size 1G (both in Spec and Status)
CSI CreateSnapshot creates a Snapshot CR
the Snapshot reconciler calls blobmgr.SnapshotBlob(, volume)
 - this calls createThinLv(, 1G) request a 1G snapshot
 - and then returns WatchPending to wait for ThinPoolLv.Status to match
CSI ControllerExpandVolume expands the Volume to 2G
the Volume reconciler calls blobmgr.CreateBlob(, 2G)
 - this updates ThinPoolLv.Spec to request a 2G volume
 - but volume.status is still 1G and we return WatchPending
the Snapshot reconciler calls blobmgr.SnapshotBlob(, volume)
 - this now calls createThinLv(, 2G)

Who wins?  We have no control whether the lvcreate will happen at a
point where volume was still 1G, or when it was 2G.  If we insist on
an exact match of the snapshot's Spec.Size to Status.Size, we may
never complete.  But since we can't control the size anyways, we might
as well just always pass 0 size down to ThinPoolLv for readonly blobs.

Hence, half the real complexity of this patch is making sure snapshots
don't break.  The rest is making sure that a requested size change
does not cause needless client.Update() calls.

Note that by itself, this patch can't be tested; it is not until a
later patch allows CreateBlob to be called while
Conditions["Available"] is true that this will matter.

Signed-off-by: Eric Blake <eblake@redhat.com>