thinpool: Update oldThinPoolLv on client.Update()
We have a couple of places where a reconciler makes more than one call
to UpdateThinPoolLv().  If the first call makes any change to state,
then we should update oldThinPoolLv to that state so that the caller
does not have to redo client.Get, yet the second call will still be
making the correct comparison on whether a change is needed.

Signed-off-by: Eric Blake <eblake@redhat.com>