thinpool: Insist that ThinLVs be empty before pool deletion
If an admin uses `kubectl delete -n kubesan-system tp $pool` while the
pool still contains LVs, we STILL want to keep the pool around for as
long as any dependent Volume or Snapshot objects have not yet finished
being deleted, so that we aren't yanking the rug out of underneath
those resources when they might be in use.  While lvm will generally
fail to delete an active pool, and fail to deactivate an in-use pool,
relying on that is riskier and makes for noisier logs and more lock
contention on lvm.

Rather than open-coding LV deletion in the cluster controller, it is
safer to just insist that we keep our finalizer and mostly ignore
DeletionTimestamp until the pool is empty.  Meanwhile, the node
controller must NOT special case DeletionTimestamp except when the
pool is empty, so that in-use volumes and the safe deletion handshake
can continue unimpeded.  As an exception to ignoring the flag, we can
insist that we will not create any new ThinLVs in a pool that has been
marked from deletion (ie. a user attempting to clone from a volume
where the pool is marked from deletion should be treated the same as
if that source is already deleted, rather than adding a new temporary
clone-lv into the source pool).

Signed-off-by: Eric Blake <eblake@redhat.com>