maint: Set config.Domain from environment
I'm working on the ability to install KubeSAN under an alternative
Group name (also called the Domain), so that we can have a single
cluster with both CRDs tied to both a release version of KubeSAN
(kubesan.gitlab.io) and a development version
(devel.kubesan.gitlab.io).  That way, developers can change the devel
CRD in incompatible ways without worrying about version-bumping the
stable CRD for cross-version compatibility, while testing things out.

For this to work, all spots in our code that need to refer to the
current choice of Group should use the same string.  This patch
focsuses on the Go and populating things from the environment set in
the deployment pods; the remaining work will be a kustomization
override for all the deployment yaml (that is, we WANT kubebuilder to
continue to target just our canonical name and the kubesan-system
namespace).  Since Go does not allow circular imports, that means
v1alpha1 cannot depend on config.Domain, so it instead becomes the new
source of truth for the codebase.

Signed-off-by: Eric Blake <eblake@redhat.com>