deploy: Add new deploy/devel setup for OpenShift
I've got access to a cluster that wants to use a stable version of
KubeSAN for heavy regression testing on one large LUN, but
simultaneously allow me development access to a smaller LUN via a
second namespace.  However, to install two versions of KubeSAN in
parallel, with API breaks in the development version where we aren't
yet worried about API compatibility in the CRD, requires installing a
separate set of CRDs under a different Group.

The previous patches made it easy to ensure the entire container is
using the same group, but now it is time to create a deployment that
specifies that group everywhere.

The transformations have to do several things:

- Update the group name.  Everywhere that a normal deployment uses
  "kubesan.gitlab.io", the new deployment uses
  "devel.kubesan.gitlab.io".  This includes replacements that have to
  do subfield replacements (such as the substring in the OpenShift
  SecurityContextConstraints.users list), as well as passing in the
  new group name to the environment of the pods.

- Use a distinct namespace.  First, this deployment creates a
  kubesan-devel namespace; then everywhere that a normal OpenShift
  deployment constrains things to the pre-existing openshift-storage
  namespace, this deployment uses kubesan-devel.

- Disambiguate cluster-wide resources.  Anything that is deployed with
  cluster visibility needs a new name to not conflict with the
  counterpart already in the cluster from a normal deployment.  For
  CRDs and CSIDriver, this is already taken care of by the group name
  updates, but for ClusterRole, ClusterRoleBinding, and
  SecurityContextConstraints, this requires use of the name
  transformer to add a 'devel-' prefix (it was nicer adding the prefix
  only to these kinds, rather than globally to all resources in the
  deployment)

In addition to deploying from this directory, you will want to create
one (or more) StorageClass objects pointing to a new VG, as well as
creating a default VolumeSnapshotClass for the new CSIDriver.  And
until CDI catches up to KubeSAN's capability, you also have to patch
the StorageProfiles.

Signed-off-by: Eric Blake <eblake@redhat.com>