rbac: Slightly tighter permissions on kubesan access
Our plugin serviceaccounts don't need to access any cluster-wide
objects - they ONLY need to communicate with instances of our own CRDs
with those instances living in the same namespace as the pods running
under the csi serviceaccounts.  Therefore, we can reduce the scope of
those permissions from ClusterRole to the narrower namespaced Role.

The CSI sidecars, on the other hand, ARE accessing cluster-wide
resources (for example, PVCs from any namespace in order to make the
right CSI gRPC calls), so those remain ClusterRoles.

Meanwhile, the role for cluster manager leader election was written
without a namespace, which would put it in the default namespace
instead of the intended kubesan-system namespace.  We didn't notice it
because the OpenShift kustomization fixed that by applying the
namespace everywhere, while the kcli test cluster isn't relying on
RBAC so it didn't matter if the role could not be found.  That, and we
haven't enabled leader election yet.

Signed-off-by: Eric Blake <eblake@redhat.com>