deploy: Allow other namespaces to view KubeSAN resources
We explicitly do not want to grant other normal k8s users the ability
to muck around with KubeSAN CRs outside of CSI interfaces (other than
the fact that kubernetes already grants the admin superuser the right
to view and edit any resource in any namespace).  However, the ability
to view our objects is something that could prove useful for
debugging, and may help kubevirt make decisions with more information
than what CSI can return to populate in the PVCs.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/114
Signed-off-by: Eric Blake <eblake@redhat.com>