csi: Restrict client to kubesan namespace
We intend to create and manage all internal CR objects in the kubesan
namespace.  Thus, there's no need to ask for kubesan resources that have
been created in any other namespace; this in turn makes listing of
objects slightly faster, and will permit the next patch which tightens
the scope of the RBAC.

I first noticed this when applying the RBAC patch without this one,
when ListVolumes and ListSnapshots started failing RBAC.  I thought
about modifying those calls to do explicit filtering, then realized
that pre-filtering is even easier; however, it was worth cleaning up
the code in the process, and adding an idea for a future optimization
of ListSnapshots.

Signed-off-by: Eric Blake <eblake@redhat.com>