tests: Now that csi-sanity passes, hard-fail on regressions
Lengthen the timeout to 15 minutes (I'm regularly seeing runs on my
laptop take more than 6 minutes, and if CI is ever under a heavy load,
that could go even slower).  Then change the test to no longer skip if
any of the csi-sanity units fail, because we are at a point where we
have all features working at a minimum level and don't want to regress
as we improve them.  This fixes a long-standing TODO from when I first
added the test.

Fixes: 2de4f4b1 ("tests: start running csi-sanity", v0.5.0)
Signed-off-by: Eric Blake <eblake@redhat.com>