csi: Add new wipePolicy:UnsafeFast parameter to StorageClass
While we must default to safe, there are clusters where defaulting to
fast is permissible.  Add a knob to allow the user to opt-in to not
wiping the contents of a Linear volume beyond the first 4M extent, for
users that are okay with the security risk.  Furthermore, it is always
safe to skip wiping on the tail end of a block that will only be used
for filesystems (a pod cannot see the uninitialized space behind a
filesystem).

We may add other wipe policies in the future, such as a wipe on delete
(if the VG starts clean, and we always wipe on delete, then it is safe
to skip wiping on creation -that slows down deletion of even thin
pools, unlike this patch, but slow deletion is less of an issue than
slow creation).

Partially fixes: https://gitlab.com/kubesan/kubesan/-/issues/101
Signed-off-by: Eric Blake <eblake@redhat.com>