thinpool: Implement volume resize
When the ThinLvSpec size is set to something larger, the node
reconciler for the active node should call lvextend (our idempotency
wrapper already handles when the requested size is not an increase
after all).  Thankfully, lvm defaults to not touching any embedded
filesystem during expansion (newer lvm has '--fs ignore' to make this
explicit, but at least minikube's lvm is too old).

We may someday want to use '--resizefs' when the Volume is used as a
filesystem rather than a block, but that does not impact this patch,
since right now, a ThinPoolLv does not know whether the upper layer is
using a block or filesystem, and we plan to chicken out and forbid
filesystem expansion at the gRPC layer until we have thought through
all the gotchas.

Just as creating snapshots is made possible through automatic resize
of the thin pool, so is automatic extension.  A guest pod that quickly
writes into all available space of the newly-enlarged block may hit
stalls while the host is busy handling the resize event, but dmeventd
gives up to 60 seconds for the additional space to become present
before the guest actually loses any I/O.

Until later patches teach Volume to adjust ThinPoolLv.Spec.SizeBytes,
the only way to test this patch for now is by manual `kubectl edit -n
kubesan-system thinpool $pool`.

Signed-off-by: Eric Blake <eblake@redhat.com>