volume: Let CreateBlob() expand linear volumes
For linear volumes, we want CreateBlob() to also handle image
expansion (as mentioned in the TODO comment that this patch resolves).
Once we expand a block volume, we need to zero the tail of the image,
and the best way to do that is to reuse the code already present for
zeroing on fresh creation.  However, tracking a mere bool of
kubesan.gitlab.io/zeroed is insufficient to zero only part of the
image, so it is time to swap over to a new counter attribute
kubesan.gitlab.io/safe-offset using the monotonic increasing counter
of the earlier patch "commands: Add support for key=counter atop lvm
tags".

Note that by itself, this patch can't be tested; it is not until a
later patch updates Conditions["Available"] to false in order to call
CreateBlob for a resize that this will matter.

NOTE: This is (yet another) incompatible change to v0.9.0.  If you
have a linear volume created by earlier KubeSAN (which only set the
old attribute), and try to expand it (which only checks the new
attribute), you will end up wiping the linear volume.  But we are
still alpha, and we are explicitly documenting that in-place upgrades
from v0.9.0 are not supported.

Signed-off-by: Eric Blake <eblake@redhat.com>