deploy: Omit version label from selectors
Deployment and DaemonSet selectors are immutable once created; any
label we stick in the selector MUST be long-term stable.  Our version
number, on the other hand, is something that we regularly expect to
bump (and once we hit v1.0, bump it with backwards-compatibility in
mind for in-place upgrades).  While trying to test another patch on a
cluster that had installed v0.10.0, I found I had to completely delete
the kubesan deployments and daemonsets before installing the patched
instance, and this patch makes that step unnecessary for releases
after v0.10.0.

Fixes: 9bcdae76 ("deploy: Use better labels on all KubeSAN resources", v0.9.0)
Signed-off-by: Eric Blake <eblake@redhat.com>