deploy: Provide VolumeSnapshotClass in deployment
For StorageClass resources, we expect the admin to create one or more
on their own: in part because we don't know the VG name(s) until the
admin installs things, and in part because we don't know if the admin
will prefer Thin or Linear volumes or allowing both.  But for
VolumeSnapshotClass resources, we only ever expect one, because we
don't have any parameters tweaking behavior of our snapshots.  Since
it is thus unchanging, we might as well install it as part of our
deployment, and insist that the VolumeSnapshotClass CRD is
pre-installed.  Since it is a global resource, it is also worth naming
it to match our actual Group name (so that parallel installs can
direct to the right parallel driver).

Signed-off-by: Eric Blake <eblake@redhat.com>