docs: Better table of v0.10.0 features
We did get Thin cloning working before v0.10.0; update that in the
documentation, along with an example of creating a snapshot.  Also, it
may be fairly easy to filesystem cloning and creation by contents (we
would clone the underlying block, and we may want to do some minor
tweaking to munge the filesystem's uuid to be unique upon
rehydrating), so we can mark that as a planned feature.

Fixes: 380a2cc4 ("docs: Recommend cloneStrategy: csi-clone for OpenShift Virtualization", v0.10.0)
Signed-off-by: Eric Blake <eblake@redhat.com>