tests: Deploy snapshot CRDs prior to creating cache
Trying to install CRDs and deploy an instance described by that CRD in
the same kustomization run does not work reliably (even if kustomize
orders things topologically, the bulk nature of multiple resources in
one go implies that the server is likely to see a request to define a
resource where it can't yet look up the CRD from cache), so the
recommendation is to do two-stage installs with the first deploy
getting CRDs in place before the second deploy creates instances of
that.

What's more, there is no reason that we should ever not support volume
snapshots.  If we're going to insist that all clusters that use
KubeSAN are already set up for snapshots, we might as well insist on
that being the case prior to the point where we cut a cache snapshot
in our testsuite, so that we aren't hitting the internet to
re-download the same CRD kustomizations for every re-spawn of the
cached kcli cluster.  I have seen Jenkins hit spurious failures more
than once when it timed out trying to download snapshot CRDs.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/108
Signed-off-by: Eric Blake <eblake@redhat.com>